Adds the bash file and some modification in the core file.
The bash file executes the ffmpeg command (Creating the images from the video). Afterwards the
cpp-code is executed. Whenever there is a false image it is put in the 'false' folder. Then ffmpeg
builds the new video from the images (now named "new%d.bmp") to a video. Finally, all images are
removed.